targets/tkey: fix implicit integer conversion warnings
Add explicit casts to silence implicit conversion warnings across
qsort.c, strlen.c, and stdio.c. Fix undefined behaviour in SWAPINIT
pointer-to-integer conversion and missing parentheses in min() macro.
Remove unused token parameter from _lengthModifier() in printf-emb.c.

+++ targets/tkey/libc/newlib/libc/string/strlen.c
@@ -35,5 +35,5 @@ strlen (const char *str)
 
   while (*str)
     str++;
-  return str - start;
+  return (size_t)(str - start);
 }
